split up widget file
