+++ src/js/janki/src/models/index.ts
@@ -1,0 +1,4 @@
+// Copyright (c) 2021 University System of Georgia and janki contributors
+// Distributed under the terms of the BSD-3-Clause License.
+
+export * from './collection';
