+++ src/js/janki/src/widgets/widgets.tsx
@@ -1,0 +1,2 @@
+// Copyright (c) 2021 University System of Georgia and janki contributors
+// Distributed under the terms of the BSD-3-Clause License.
